renamed ipa_canopen_chain_ros to canopen_chain_node
