+++ canopen_chain_node/src/chain_node.cpp
@@ -1,12 +1,12 @@
 #include <socketcan_interface/dispatcher.h>
 #include <socketcan_interface/socketcan.h>
-#include <ipa_canopen_chain_ros/chain_ros.h>
+#include <canopen_chain_node/chain_ros.h>
 
 using namespace ipa_can;
 using namespace ipa_canopen;
 
 int main(int argc, char** argv){
-  ros::init(argc, argv, "ipa_canopen_chain_ros_node");
+  ros::init(argc, argv, "canopen_chain_node_node");
   ros::NodeHandle nh;
   ros::NodeHandle nh_priv("~");
